# swagger.json
{
    "openapi": "3.0.2",
    "info": {
      "title": "Staff API",
      "version": "1.0",
      "description": "An API for managing staff, customers, and transactions."
    },
    "paths": {
      "/login/login": {
        "post": {
          "tags": ["login"],
          "summary": "User Login",
          "operationId": "userLogin",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Login"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Login successful",
              "content": {
                "application/json": {
                  "example": {
                    "message": "Login successful"
                  }
                }
              }
            },
            "401": {
              "description": "Invalid credentials"
            }
          }
        }
      },
...
      "/login/register": {
        "post": {
          "tags": ["login"],
          "summary": "User Registration",
          "operationId": "userRegistration",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Register"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Registration successful, check email for verification",
              "content": {
                "application/json": {
                  "example": {
                    "message": "Registration successful, check email for verification"
                  }
                }
              }
            },
            "400": {
              "description": "Email or Username already taken"
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
        "Login": {
          "type": "object",
          "properties": {
            "email": {
              "type": "string"
            },
            "password": {
              "type": "string"
            }
          },
          "required": ["email", "password"]
        },
        "Register": {
          "type": "object",
          "properties": {
            "username": {
              "type": "string"
            },
            "email": {
              "type": "string"
            },
            "password": {
              "type": "string"
            }
          },
          "required": ["username", "email", "password"]
        },
...
      }
    }
  }